% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dough.R
\name{build_recipe}
\alias{build_recipe}
\title{Build a sourdough recipe from final weight}
\usage{
build_recipe(
  final_weight,
  hydration = 0.7,
  pct_starter = 0.25,
  starter_hydration = 1,
  pct_salt = 0.02
)
}
\arguments{
\item{final_weight}{Numeric. Target weight for dough, in grams. This won't be
completely accurate because salt is added to the final mass. See Details
for more information.}

\item{hydration}{Numeric. Hydration for dough, in percent. Defaults to 0.7 (70\%).}

\item{pct_starter}{Numeric. Percentage of starter for dough. Defaults to 0.25 (25\%).}

\item{starter_hydration}{Numeric. Level of hydration of starter. Defaults to 1 (100\%).}

\item{pct_salt}{Numeric. Percentage of salt for dough. Defaults to 0.02 (2\%)}
}
\value{
Text printed to the console with a recipe.
}
\description{
Generate a recipe based on the final target weight of the loaf dough.
}
\details{
Salt is currently just added on as extra because I couldn't figure out the
math for it :shrug: (i.e. specifying a 800g final loaf will result in 809g
with 2\% salt because it just adds it to the 800g)
}
\examples{
build_recipe(final_weight = 1000)

build_recipe(final_weight = 880,
                    hydration = 0.74,
                    pct_starter = 0.2,
                    starter_hydration = 1,
                    pct_salt = 0.02)
}
