% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dough.R
\name{calculate_recipe}
\alias{calculate_recipe}
\title{Calculate the hydration of a given recipe}
\usage{
calculate_recipe(flour, water, starter, starter_hydration, salt)
}
\arguments{
\item{flour}{Numeric. Amount of flour, in grams.}

\item{water}{Numeric. Amount of water, in grams.}

\item{starter}{Numeric. Amount of starter, in grams.}

\item{starter_hydration}{Numeric. Level of hydration of starter, in percent
(1 = 100\%)}

\item{salt}{Numeric. Amount of salt, in grams.}
}
\value{
Text printed to the console with a recipe.
}
\description{
Determine the hydration of a recipe, given its proportions of water, flour,
starter, and salt.
}
\examples{
calculate_recipe(flour = 450,
                 water = 320,
                 starter = 100,
                 starter_hydration = 1,
                 salt = 8)
}
