% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starter.R
\name{feeding_from_ratio}
\alias{feeding_from_ratio}
\title{Determine starter feeding amounts based on ratios}
\usage{
feeding_from_ratio(starter = 1, flour = 4, water = 4, flour_g = 100)
}
\arguments{
\item{starter}{Numeric. Ratio of starter. (e.g., 1 in a 1:4:4 ratio.)}

\item{flour}{Numeric. Ratio of flour (e.g., 4 in a 1:4:4 ratio.)}

\item{water}{Numeric. Ratio of water (e.g., 4 in a 1:4:4 ratio.)}

\item{flour_g}{Amount of flour to use in the feeding}
}
\value{
Text printed to the console with feeding amounts.
}
\description{
Generate measurements for feeding a starter based on starter:flour:water ratios like 1:4:4, 1:5:5, etc.
}
\examples{
feeding_from_ratio()

# A 1:5:5 feeding with 50g of flour
feeding_from_ratio(
  starter = 1, flour = 5, water = 5,
  flour_g = 50
)
}
